+dimoo

+++ divisor/dimoo_modules/generators/text_understanding_generator.py
@@ -70,7 +70,7 @@ def generate_text_understanding(
             x0 = torch.argmax(logits_with_noise, dim=-1)  # b, l
 
             if remasking == "low_confidence":
-                p = F.softmax(logits.to(torch.float64), dim=-1)
+                p = F.softmax(logits.to(torch.bfloat16), dim=-1)
                 x0_p = torch.squeeze(torch.gather(p, dim=-1, index=torch.unsqueeze(x0, -1)), -1)  # b, l
             elif remasking == "random":
                 x0_p = torch.rand((x0.shape[0], x0.shape[1]), device=x0.device)
